Added js files to modules
